[server] bct and sst don't accept invalid arguments

+++ src/server/commands/graphic_commands/sst.c
@@ -11,19 +11,19 @@ uint8_t srv_cmd_sst(server_config_t *server,
 	server_user_t *user, cmdparams_t *cmd)
 {
 	char *msg;
-	unsigned int frequency;
+	int frequency;
 
 	if (cmd->nb_args < 1) {
 		dprintf(user->fd, "sbp\n");
 		return (1);
 	}
-	frequency = (uint8_t) atoi(cmd->args[0]);
+	frequency = atoi(cmd->args[0]);
 	if (frequency <= 0) {
 		dprintf(user->fd, "sbp\n");
 		return (1);
 	}
-	server->frequency = frequency;
-	asprintf(&msg, "sst %u\n", frequency);
+	server->frequency = (unsigned int)frequency;
+	asprintf(&msg, "sst %d\n", frequency);
 	send_msg_to_all_graphic(server, msg);
 	free(msg);
 	return (0);
